pyproject.toml
[project]
name = "codeowner-generator"
version = "0.1.0"
description = "Generate GitLab CODEOWNERS file by analyzing git repository history"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.40",
    "click>=8.1.7",
    "python-dateutil>=2.8.2",
]

[project.scripts]
codeowner-generator = "codeowner_generator.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
